web_blog/web_blog/urls.py
"""web_blog URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/2.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, re_path, include
from django.contrib.auth import views
from blog.views import PostListView
from restapp.views import UserViewSet
from albums.views import AlbumViewSet, ArtistViewSet
from rest_framework import routers

# Routers provide an easy way of automatically determining the URL conf.
router = routers.DefaultRouter()
router.register(r'users', UserViewSet)
router.register(r'albums', AlbumViewSet)
router.register(r'artists', ArtistViewSet)


urlpatterns = [
    path('admin/', admin.site.urls),
    path('index', PostListView.as_view(), name='index'),
    path('', include('blog.urls')),
    path('accounts/login/', views.LoginView.as_view(template_name='registration/login.html'), name='login'),
    path('accounts/logout/', views.LogoutView.as_view(), name='logout', kwargs={'next_page': '/'}), # redirecting to '/' will take you back to the homepage or regex''
    path('test/', include('apppass.urls')),
    path('mpform/', include('formwizzard.urls')),
    path('rest/', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    path('api/', include(router.urls)),
    path('datatables/', include('albums.urls'))
]
