web_blog/requirements.txt
Django==2.2.4
django-formtools==2.1
djangorestframework==3.10.2
djangorestframework-datatables==0.5.0
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
